refactor(node): reuse callback wrapper factories
